Added new edit_item_page to allow more field editing inc description and title. General html format improvements. Moved class defs into trello_items to preserve naming style with old 'session_items.py' file. Tidied up project files
