--- todo_app/.gitignore
@@ -1,1 +1,0 @@
-USER_INFO
