+++ todo_app/templates/index.html
@@ -36,7 +36,7 @@
 
   <div class="jumbotron">
     <h1 class="bigTitle"><b>[to_do||!to_do]</b></h1> 
-    <p style="margin-left: 30em; color: #333333;"> Version: {{ version }} </p>
+ <!--   <p style="margin-left: 30em; color: #333333;"> Version: {{ version }} </p> -->
     <p class="subTitle"> &nbsp; ...that is the question...</p>
   </div>
 
@@ -61,16 +61,20 @@
         <!-- Use a for loop to process all the items -->
         {% for item in todo_items %}
           <p style="color:#e0e0e0">
-            <b>{{ item.title }} &nbsp; | &nbsp; </b>    
             {% if item.list['name'] == 'To Do' %}
               <b style="color:#ec6161">{{ item.list['name'] }}</b> <b> &nbsp; | &nbsp; </b> 
+              <b>{{ item.title }} &nbsp; | &nbsp; </b>    
               <a href="{{ url_for('update_status', id = item.id, new_status = 'Doing') }}">Start Task</a> <b> &nbsp; | &nbsp; </b>
             {% elif item.list['name'] == 'Doing' %}
               <b style="color:#f3e566">{{ item.list['name'] }}</b> <b> &nbsp; | &nbsp; </b>
+              <b>{{ item.title }} &nbsp; | &nbsp; </b>    
               <a href="{{ url_for('update_status', id = item.id, new_status = 'Done') }}">Complete Task</a> <b> &nbsp; | &nbsp; </b>
             {% elif item.list['name'] == 'Done' %}
               <b style="color:#65ec61">{{ item.list['name'] }}</b> <b>&nbsp; | &nbsp; </b>
+              <b>{{ item.title }} &nbsp; | &nbsp; </b>    
+              <a href="{{ url_for('update_status', id = item.id, new_status = 'Doing') }}">Re-open Item</a> <b>&nbsp; | &nbsp; </b>
             {% endif %}
+            <a href="{{ url_for('edit_item_page', id = item.id) }}">Edit Item</a> <b>&nbsp; | &nbsp; </b>
             <a href="{{ url_for('clear_item', id = item.id) }}">Clear Item</a>
           </p>
         {% endfor %}
